requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.2
allure-python-commons
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
configobj==5.0.8
crcmod==1.7
cryptography==41.0.3
curlify==2.2.1
exceptiongroup==1.1.3
Faker==19.6.1
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
jsonpath==0.82.2
oss2==2.18.1
packaging==23.1
pluggy==1.3.0
pycparser==2.21
pycryptodome==3.18.0
pyotp==2.9.0
pytest==7.4.2
pytest-assume==2.4.3
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tomli==2.0.1
urllib3==2.0.4
ruamel.yaml==0.17.29
pytest-html
pytest-rerunfailures
Allure-pytest
loguru~=0.7.0
requests_toolbelt
pytest-dependency
jmespath